Describe "via" model in many2many fields.
It is sometimes impossible to determine which table is used for the actual references in many2many relationships, because while some use the canonical `these_those_rel` naming scheme, not all do, and even the ones that do don't specify the order of `these` and `those`.  This just adds 3 more `_description_`s to the `Many2many` class - `via_relation`, `via_these`, and `via_those` so that the `get_fields` API call can know all about how those relationships work.